Create IsolateServer in both run_isolated.py and isolateserver.py.
This is to eventually merge the code with isolateserver.py but this is a
stepping stone so the CL is not too large. In the meantime it is a bit weird
because there is two IsolateServer class but the class will eventually be moved
into isolateserver.py, which run_isolated.py will import.

Remove --download support from run_isolated.py. It will be added back as
isolateserver.py download subcommand later.

R=vadimsh@chromium.org
BUG=

Review URL: https://chromiumcodereview.appspot.com/23851024