Fix about:tracing import errors with cc::Tiles
Tiles can be used in either cc.debug or cc.debug.quads. Update their category
accordingly. Tracing importers currently require that the category for a tile
when it is snapshotted match the category when it is deleted. We might relax
that in the long run, but thats anther discussion for another time.

R=danakj@chromium.org
CC=danakj

Review URL: https://codereview.chromium.org/23461047