Fix Views Textfield keyboard context menu placement.
Show Textfield key-triggered context menus at the text cursor.
(previously it showed the menu in the middle of the textfield)
It now anchors at the bottom and to the right of the caret.
(Win textfields show at top right... but that covers text)
(It's mirrored well in RTL UI and accounts for screen edges)

Pipe Views value through the textfield wrapper to make this happen.
(Win impl works as-is and is not affected, soon to be removed)
I split this from http://crrev.com/222434 for merge consideration.

BUG=276822
TEST=Menu and Shift+F10 shows textfield context menus at the caret.
R=sky@chromium.org

Review URL: https://chromiumcodereview.appspot.com/24097003