Don't flash black in classic, or in the body in aero
Maintains the clear in the title bar so that DWM doesn't flash white in the
non-client area. Don't clear the main content area despite the drawbacks for
resize on Win8 because the black clear on a new window is too noticable
because the NTP has a white background.

TEST=new windows, new menus, and resizing on win7 classic, aero, and win8
R=sky@chromium.org
BUG=177115,289095,290481

Review URL: https://chromiumcodereview.appspot.com/23842009