[Android WebView] Wire up Blink's WideViewportQuirkEnabled setting
This setting should be always enabled by now, independently of other
quirks. Added tests to ensure that UseWideViewport works regardless
of whether quirks are enabled or not.

The WideViewport quirk is to maintain compatibility with Android apps,
and is required so long as WebSettings.{get|set}UseWideViewPort is
supported.

Also enabled AwSettingsTest#testUseWideViewportLayoutWidth since the
fix in Blink has been rolled in.

BUG=285995,288037

Review URL: https://chromiumcodereview.appspot.com/23496047