CC: Add a scheduled action for ManageTiles
Other than commits, ManageTiles should be coalesced to
happen on a semi-regular basis. This used to happen during
draws, but this is not reliable as draws might not occur,
nor we shouldn't force them to.

This adds a scheduled action for ManagingTiles, and the
ability to request this action, similar to requesting draws
etc.

BUG=264566

Review URL: https://chromiumcodereview.appspot.com/23495022