Fix the build target on Mac10.8 Tests so that we build a full archive.
We were only building the targets we needed for testing, but now that
we're archiving the builds so that we can bisect Blink changes for QA,
we need to build everything that the archive step is looking for.
We already  do that on Linux and Win, so this just brings Mac up to match them.

R=iannucci@chromium.org
BUG=266421

Review URL: https://chromiumcodereview.appspot.com/23710056