Replace media::VideoCapture::VideoFrameBuffer with media::VideoFrame.
This CL removes a redundant class in the form of
media::VideoCapture::VideoFrameBuffer, and replaces it with
media::VideoFrame.  Also: as media::VideoFrame has an embedded
destruction callback, use this to handle buffer returns on
the video capture stack, and remove the explicit
VideoCapture::FeedBuffer entry point.

BUG=None
TEST=local build, content_unittests, run on desktop Linux with
     screen capture and webcam

Review URL: https://chromiumcodereview.appspot.com/23587018