Create the Tooltip widget in Aura as needed instead of just creating it in the timer.
Currently the Tooltip widget is created in the tooltip timer which also handles setting the tooltip text on
the Tooltip object. There is a timing window between this and the mouse event coming in, which ends up causing
the tooltips to not show up if the underlying root changes in between.

Fixes bug https://code.google.com/p/chromium/issues/detail?id=292530

BUG=292530
R=sky@chromium.org, sky

Review URL: https://codereview.chromium.org/23958012