Bug fix: Append a language to the list after blocking it for Translate on Chrome OS
Chrome OS and the other platforms are different in terms of the preference for the language list of chrome://settings/languages, so I factored the functions to manipulate the list and have it called when blocking the language for Translate.

NOTE: Removed DCHECK(...) in translate_infobar_delegate.cc to check if the language is listed in the blocked language list here because the language might already be listed. For example:
1. After the user add the language, he/she removes the language chrome://settings/languages. In this case, Translate infobar will appear.
2. While the Translate infobar is shown, the user can change if Translate should be offered in the language. In this case, the user can push Never Translate (language)" button while the language is already blocked.

BUG=285651

Review URL: https://chromiumcodereview.appspot.com/23923007