Files.app: Added the nameText field to TestEntryInfo class.
Originally, we assume the target name of entry provided to FakeFileSystem and
expected name actually shown in the file list is the same. But for hosted
documents on the drive, both name can be different. To handle the case, this CL
adds the nameText field to TestEntryInfo class.

BUG=279774
TEST=file_manager_browsertests

Review URL: https://chromiumcodereview.appspot.com/23926014