Files.app: Move the error handler to the top of combined script.
Currently the error handler to count uncaught errors is set in the file_manager.js.
This CL moved the error handler to the top of combined script, which enable us to catch the uncaught erros that is raised before the file_manager.js is loaded.

BUG=none
TEST=manually

Review URL: https://chromiumcodereview.appspot.com/23536047