This changes the current directory for script execution to be the build directory rather than the current BUILDfile directory. This makes things more consistent and simplifies some implementation.
It adds a new function to_build_path() for buildfiles to convert files to the build directory for this kind of call.

I removed all of the built-in relative* variables that were required to support this, and added one for target_out_dir and root_out_dir.

I added some tests and a test helper object since I was doing this kind of Scope setup a lot.

R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/23606031