chromeos: Fix display failures when going to mirrored mode.
This fixes an issue where mirrored mode configuration would
fail when the internal and external displays didn't share a
common resolution. A new mode would be created on the
internal display, but it wasn't added to the corresponding
OutputSnapshot struct, resulting in a later failure.

It also moves the mirror-mode-finding code from
RealOutputConfiguratorDelegate to OutputConfigurator so it
can be tested.

BUG=289159

Review URL: https://chromiumcodereview.appspot.com/24081004