Initial version of the "NaCl and PNaCl" comparison document.
The idea is that the "default" documentation flow presented by the dev site to new developers will describe PNaCl. NaCl *is* mentioned in the Technical Overview but this document is going to be linked for those needing more information.

----

Should this be part of the technical overview instead of a separate doc?

Pros: Some of the content overlaps. The technical overview doesn't dedicate
as many words for describing the differences and the contents listed here could
be useful (as well as remove a small amount of natural duplication between the
two documents).

Cons: From the point of view of developers new to the technology, most don't
and should not care about this document. The technical overview, OTOH, is an
important initial gateway to Native Client and thus it should stay as simple
as possible, as well as simply describe PNaCl as the default toolchain.

For now leaving it separate: can be merged later.

BUG=None
R=binji@chromium.org, jfb@chromium.org, mackinlay@google.com, sehr@chromium.org

Review URL: https://codereview.chromium.org/23766016