Docserver: Rewrite GithubFileSystem
To support new features in the SamplesDataSource rewrite, the
GithubFileSystem needs to be able to access arbitrary (public)
repositories, not just the chromium samples repo.

There has been a long standing TODO to have the GithubFileSystem use
a CachingFileSystem instead of handling caching logic internally.

Some other issues with the old implementation lead to a new
GithubFileSystem implementation.

However, this patch's scope is limited. The old GithubFileSystem has not
been replaced yet and the new one is in a file named 
new_github_file_system.py. The SamplesDataSource is the currently the 
only consumer of a GithubFileSystem, and because it is being rewritten
anyway, I don't want to waste effort time making the new file system
work with the old data source.

Replacing the old file system and integrating the new one will be done
in the samples data source rewrite patch.

NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/23454026