Add zero-delay polling option to sensor data fetcher.
Add the option to configure the poll delay of the polling data fetcher
thread. If the delay is zero the thread is started but does no
polling by itself, i.e. Fetch() is never called. The latter option is
useful when we want to create any platform sensor components on a
separate thread, but don't require the Fetch() callback method.

BUG=261165

Review URL: https://chromiumcodereview.appspot.com/24125003