Add an assertion to prevent mapping connection failures to ERR_IO_PENDING.
This could happen if connect() were to set the error as WSA_IO_PENDING. Although this isn't a documented error on MSDN, there is some crash evidence suggesting it is happening.

The code does not expect WSA_IO_PENDING and this will lead to a use-after-free.

BUG=244727

Review URL: https://chromiumcodereview.appspot.com/24175006