Remove SendFakeProcessKeyEvent.
There is a weired hack in UpdatePreedit or CommitText which sends
VKEY_PROCESS(229) into to text input client if the event is not based on keyboard
event. However sending VKEY_PROCESS key event means that some key is consumed by
input method. So that it is not appropriate to generate.

BUG=289105
TEST=Manually done

Review URL: https://chromiumcodereview.appspot.com/24123006