Fix master name for uploading test results from android bots.
We need to pass the value "ChromiumWebKit" to test-results.appspot.com,
not "chromium.webkit". The non-Android bots get this string from
chromium_utils.GetActiveMaster(), but it's not obvious how to get this
in the build/android/buildot code, so I'm hard-coding it for now.

TBR=ilevy@chromium.org
BUG=293928

Review URL: https://codereview.chromium.org/23496068