For App List Search, update apps on uninstall, rather than unload.
As of http://crrev.com/218760, terminated and disabled apps are included
in the app list search. Now, an uninstall is a better signal to update
the list, because the unload notification is not sent out for disabled
apps.

BUG=292985
TEST=In chrome://extensions disable an app by unticking the checkbox.
Then, in the app launcher, uninstall the greyed-out app from the grid.
Then, try searching for it in the app launcher. The app should not
appear, and Chrome should not crash.

Review URL: https://chromiumcodereview.appspot.com/23526059