[NaCl SDK] nacl_io: make remove(3) glibc-only
Both newlib implements remove in terms of unlink(2) and
rmdir(2) so there is no need to intercept it at this level.
What is more it is causing link errors in certain
cucumstances because newlib's remove lives in the same
objects _remove_r.

Ideally we would remove this function for glibc as well but
the plumbing is not yet in place for this.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/23588004