Use specified cache size for PNaCl translation cache unit tests
The default implementation uses the disk cache's size heuristic which
depends on the available disk space. If one of the bots happens to be
short on disk space and gets a small cache, this test can fail.
This CL simplifies the PnaclTranslationCache init interface, and uses
the private version in the unit test, so we can control the size.
This should mean we can re-enable the large file test.

R=jvoung@chromium.org
BUG=288745
TEST= unit_tests PnaclTranslationCacheTest.*

Review URL: https://chromiumcodereview.appspot.com/24227002