sync: Gracefully handle early shutdown
Introduce a new object to communicate cross-thread cancellation signals.
This new object, the CancellationSignal, is protected by a lock. It
allows the receiving thread to query whether or not a stop has been
requested. It also allows the receiving thread to safely register a
cross-thread callback to be invoked immediately when a stop is
requested.

We use two instances of this class to ensure we meet all the
requirements for a safe and fast sync backend shutdown.

The first instance is used with the HttpBridgeFactory to allow the UI
thread to force it to drop all refereces to its RequestContextGetter
immediately.  This is an important part of our plan to ensure that all
references to that object are released before
ProfileSyncService::Shutdown() returns, which is necessary to avoid
racy crashes at shutdown.  (See crbug.com/236451)

The second instance is used with the ServerConnectionManager and the
Syncer.  Once signalled, it ensures that any active connections are
released (possibly decrementing another ref to the
RequestContextGetter), that any blocking I/O is aborted, and that no
more connections will be instantiated.  It's important to prevent the
creation of more connections because the HttpBridgeFactory might trigger
a crash if we asked it to create another connection after it had been
shut down.

The syncer's interaction with the second cancelation signal is more
passive.  It does not execute any callbacks when the signal is sent.
Instead, it queries the signal as it performs sync cycles, and will cut
short any existing sync cycle if it notices the signal has been sent.

Finally, this CL includes one important change to the initialization of
the HttpBridgeFactory.  In order to properly register with the
cancellation signal while still allowing the creation of the user agent
to occur on the sync thread, its initialization had to be split into two
parts.  This class now has an Init() method in addition to its
constructor.

BUG=236451

Review URL: https://chromiumcodereview.appspot.com/23717047