CC: Disable raster-on-demand during prefer-smoothness
We prioritize memory in the active tree during accelerated
gestures, so we can't trigger raster-on-demand if the
pending tree tiles are missing.

This patch delays raster-on-demand, by not
activating until the accelerated gesture is complete.

This patch follows these two fixes (which guarantee
activation, and showing new visible tiles while waiting
for activation).
https://codereview.chromium.org/23686011/
https://codereview.chromium.org/23495022/

BUG=264566

Review URL: https://chromiumcodereview.appspot.com/23921005