Restart faster with flags by not waiting for browser tear down.
Replaces the call to ExitCleanly with AttemptExit on the path to
restart the browser with paramters which doesn't wait for the browser
windows to report full tear down which is fine since at this point
there will be no windows to wait for. It isonly important to mark the
exit as a clean one to avoid getting the browser crashed infobar.

BUG=293466
TEST=Only kae sure that the "Browser has crashed" infobar is not present when applying flags.

Review URL: https://chromiumcodereview.appspot.com/24253003