Fixed rendering of the webui list.
Recently all of the items in the webui list got cached, for performance. However, this caused one issue. In webui's list with autoExpand option set to false, only visible items are rendered + the selected item. Therefore, if the selected item is out of the visible rect, then it will be rendered, but hidden. This is to keep focus, aria selection indexes, etc. And the problem occurs here. If the selection changes, the pinned item stays in cache with the hidden attribute on. If it is later rendered, then it is basically hidden. That's all. Therefore, we need to keep it in cache with removed the hidden attribute. That's what this patch does to resolve this issue.

The bug caused lots of problems in rendering in the Files app, like disappearing files.

TEST=Tested manually.
BUG=294504,289682,275275,partner:22244

Review URL: https://chromiumcodereview.appspot.com/24251002