Add PNACL_CACHE cache backend type, and allow larger files for this type.
PnaCl's translation cache is currently using the DISK_CACHE backend,
but we don't want to mix the UMA statistics for these caches. Also,
because we expect PNaCl's cache to have a small number of files, which
will be much larger than most HTTP cache files, we remove the restriction
that files can take up no more than 1/8 of the cache. For machines with
little available disk space (and thus small caches), this may mean the
difference between being able to store any files in the cache, or not.

R=rvargas@chromium.org,jvoung@chromium.org
BUG=none
TEST= unit_tests PnaclTranslationCacheTest.* (and no net_unittests break)

Review URL: https://chromiumcodereview.appspot.com/24208003