Modified picture raster trace to report the tile's clip rect.
Modified the rect that is passed to Picture::Raster to be the 
clip-rect corresponding to particular tile that's being drawn. This 
improves the data reported by the Raster trace event which now accurately
reports the offset and size of the tile in content space. It was previously
reporting the clip rect of the entire SkPicture (or a portion of it
when there's overlap in the pile) which wasn't that useful.

BUG= 292248

Review URL: https://chromiumcodereview.appspot.com/23876028