Don't UpdateCursorIfOverSelf when the current root window isn't focused
When the cursor gets set, this code checks to see if it's over itself
as far as aura::Window goes it is. However, when a menu pops up, this
is still true, so make sure the RootWindow for the Window that is
checking also has focus.

R=ben@chromium.org
BUG=292379

Review URL: https://chromiumcodereview.appspot.com/24244005