[Activity log] Drop XMLHttpRequest.open from the argument whitelist
Do not keep all arguments for the XMLHttpRequest.open DOM call.  Usually
we will extract the URL argument into arg_url where it is accessible
even when arguments aren't kept.  Sometimes, however, the argument
contains a relative (not absolute) URL which can't be extracted.  In
this case, err on the side of privacy and do not log the URL at
all--rather than keeping the arguments around, where the URL is kept but
not sanitized or cleared with history.

Review URL: https://chromiumcodereview.appspot.com/23928006