Recipe for the GPU bots.
This recipe is intended to cover all of the GPU bots:

  - chromium.gpu
  - chromium.gpu.fyi
  - The GPU bots on the chromium.webkit waterfall
  - The GPU try bots on the tryserver.chromium waterfall

This recipe has been run locally on Linux and is known to complete successfully there. Further testing will be done on other platforms before switching over the waterfalls.

There are some TODOs regarding how the recipe will be invoked and what it will contain, but the initial switch over to using the recipe will be done before addressing the former.

BUG=286398

Review URL: https://chromiumcodereview.appspot.com/23702029