Adding a key code map for easy keyboard event simulation.
BUG=134210

It is very tedious to simulate a key press event: one needs to find the |code| and |KeyboardCode| corresponding to the character being simulated and then call SimulateKeyPress. This change adds a key code map that translate a char to the corresponding codes assuming US keyboard layout.
This key code map is not suitable for simulating keyboard events originated from a keyboard with a non-us layout. But it is good for the majority of the test scenarios.
It makes keyboard event simulation in the chromoting browser_tests much simpler.

Review URL: https://chromiumcodereview.appspot.com/24233002