Don't show uninstall prompt in App Launcher unless caller asks.
Before this change the code will always show the uninstall prompt if the
app list is visible. This is bad for the case where the app launcher is
open and the user clicks the remove icon in chrome://extensions. Before
the app launcher has a chance to dismiss itself, it will have an
uninstall prompt shown underneath it.

This is doubly bad because the delegate will not notify the app launcher
that it has a dialog underneath it, so the app launcher can be dismissed
from behind the uninstall prompt.

BUG=292466

Review URL: https://chromiumcodereview.appspot.com/24244002