<webview>: Reduce test flake.
This CL reduces <webview> test flake by not loading the test web server if the server is not necessary for the test. It also modifies the edit_commands test so that it no longer requires a test server.

This CL also re-enables WebViewInteractiveTest.PointerLock_PointerLockLostWithFocus in the hope of having removed flake as it no longer loads a test server.

BUG=293445

Review URL: https://chromiumcodereview.appspot.com/24244004