Dispatch a ui::ET_MOUSE_EXITED event when the cursor becomes invisible
Clear existing hover effects in views when the cursor
becomes invisible by dispatching a mouse exit at
the cursor's location. Note we do not need to dispatch a
mouse enter when the cursor becomes visible because that
can only happen in response to a mouse event, which will
trigger its own mouse enter.

The early exit in Widget::OnMouseEvent() is no longer
needed since the issue it was originally meant to fix
(issue 239363) is also fixed by revision 208303. It is
being removed here since it blocks the dispatch of
mouse exit events.

BUG=244262,253317
R=oshima@chromium.org, sadrul@chromium.org

Review URL: https://codereview.chromium.org/23935007