Switched all desktop bots on the chromium.gpu waterfall to the new GPU recipe.
This must wait for https://codereview.chromium.org/23702029/ to land.
Waiting to commit that until a test run has completed on Windows.

Tested this with runbuild.py --show-commands and verified the expected
factory properties are being passed down to the annotator on each builder.

Once the chromium.gpu waterfall is working well, the GPU bots on the
other waterfalls will be switched over.

BUG=286398

Review URL: https://chromiumcodereview.appspot.com/23530069