Manually landing already reviewed https://codereview.chromium.org/23876020/ on behalf of michaelpg@:
Multiprofile avatar menu for ChromeOS users.

This change affects ChromeOS when the multiple profiles flag is set.
If multiple user accounts have been logged in on a device, the avatar menu
will include a link to add a user to the session. Once some number of users
is added, the avatar menu will display all logged in users and allow the user
to switch to, or open, a browser window for a given profile.

AvatarMenuModel moved to AvatarMenu, model and controller functions separated out into interfaces ProfileList and AvatarMenuActions. AvatarMenu instantiates components based on multi-profile flag and platform. ChromeOS version of ProfileList uses the UserManager to identify logged-in users and translate them into their corresponding profiles.

BUG=288365
TEST=manual, added /chrome/browser/chromeos/profiles/profile_list_chromeos_unittest.cc
TBR=ben,erg,dpolukhin,xiyuan,michaelpg

Review URL: https://codereview.chromium.org/23503085