EVEA: Switch to MMAP memory type for MFC output buffers.
Use MMAP buffers and copy out to shared memory from them instead of passing shared memory as USERPTR to MFC. This works around corruption in output stream and the additional memcpy is in range of around a few to a few tens of kilobytes for 720p streams. This doesn't even apply to keyframes, because we were already doing a memmove of the data in their case to put the "header" in front of the dequeued stream.

The corruption happens, because the kernel keeps the USERPTR buffers it set up even after giving them back to us and tears that setup down only if we queue a different buffer than previously under given index. But the old buffer might already be in use somewhere else and the tear down may corrupt the contents in a way that we have not fully grasped yet. Contributing to this was an implicit cache sync on tearing down, but this appears not to be the only issue, as removing it didn't fix the corruption.

This should also actually improve performance. The process of queuing a USERPTR buffer includes copying vma struct for the area, pinning all pages that belong to the area (which involves laboriously walking through the page tables, pulling each page into memory from backing store if needed and incrementing refcounts on each of them), then allocating SG table for all the pages, mapping all pages in device's IOMMU, syncing the cache on noncoherent architectures and verifying contiguity of the memory. Adding to this the fact that our output buffers are 2MB in size, and that we effectively have to, in addition to do doing all the above operations, repeat them in reverse order to clean up in current implementation on each frame, a memcpy of a hundred kilobytes or so seems quite preferable. In MMAP case we only have to sync the cache (once).

As measured by hshi@:
Under Google Cast extension's High (720p) quality setting. Test page URL: Spinning WebGL Box
(www.khronos.org/registry/webgl/sdk/demos/webkit/SpiritBox.html)
Without this CL, about 18.5fps
With this CL, about 23.3fps


BUG=289775,284783
TEST=apprtc

Review URL: https://chromiumcodereview.appspot.com/23653039