Relanding this as the previous attempt failed for ChromeFrame. That was because we were not resizing the Container window when the content window is resized. On ChromeFrame the window is initially created with a zero rect and is then resized.
Add a container window to host the desktop window hierarchy in AURA.

This ensures that ZOrder changes within this hierarchy don't affect the ZOrder of windows like tooltips
which are direct children of the root.

This change fixes the issue with the tooltips getting hidden behind the browser window during a tab drag drop
operation.

Reverted the earlier fix in tooltip_controller.cc which destroyed the tooltip when the mouse left the window
bounds. While that works this appears to be a cleaner fix based on a discussion with sky.

I disabled the PopupBlockerBrowserTest.WindowFeatures test on linux aura as it fails for some reason
with this change. Logged a bug to track that https://code.google.com/p/chromium/issues/detail?id=295299

BUG=292530
TBR=ben
TEST=Covered by new views unittest TooltipAtTopOfZOrderAfterActivation

Review URL: https://codereview.chromium.org/24343002