Make WebRTC slaves use auto_reboot default (True)
Change all machines to reboot between builds instead of just 
Windows machines. We've seen some strange errors with some tests
that will be easier to debug knowing that the machine is always
freshly rebooted.

This CL also makes sure the auto_reboot value is not overwritten
in production, in case some bots will be configured to use
auto_reboot=False in the future. In that case, the auto_reboot
attribute should be configured for each individual builder.

TEST=none
BUG=none

Review URL: https://chromiumcodereview.appspot.com/24088012