[PPAPI] ResourceVar now reference counts its Resource in the plugin.
ResourceVar is now an abstract base class with subclasses
HostResourceVar and PluginResourceVar. The PluginResourceVar has a
reference counted Resource instead of a PP_Resource.

VarTracker has MakeResourceVar and MakeResourcePPVar methods, to
abstract over the creation of a resource var of the correct subclass.

Also, the creation_message is now NULL when empty, instead of being an
empty message object.

BUG=290713

Review URL: https://chromiumcodereview.appspot.com/23809016