This CL's goal is to answer the following question: "For each pageload, how many extensions do X thing to the page's DOM?"
The new UmaPolicy class processes ActivityLog actions and selectively histograms "interesting" actions. In order to avoid duplicate counting, the UmaPolicy class keeps track of open tabs & the extensions that operate on them. It records to the histograms when a page is closed. Page are considered "closed" when the user closes a tab or navigates to a new page in the same gab.

The UmaPolicy class does not replace the existing ActivityLog database policies. Instead, they co-exist. For this reason there are some small changes in activity_log.cc|h and related tests to distinguish between the database policies and the UMA policy.

For privacy reasons, the Activity Log does not initialize the UmaPolicy for incognito profiles.

This CL also includes unit tests and browser tests for UmaPolicy.

BUG=273895
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/24212002