Add a test extension signed with a 2048-bit key
Add an extension signed with a 2048-bit key to the
chrome/test/data/extensions directory.  A unit test that checks loading
of the new extension will be checked in separately; the .crx file is
being committed separately without any test changes since it is a binary
file and can't go in via the commit queue.

The full change is being reviewed in
https://codereview.chromium.org/23531061/

BUG=293661
R=felt@chromium.org

Review URL: https://codereview.chromium.org/23494061