Safer safe-browsing chunk range decoding.
Safe browsing stores encode list id as the low-order bit of chunks.
Some stores only contain one list, but in case of corruption the
low-order bit cannot be guaranteed to be 0.  Change to decode two
lists in all cases, and discard the unwanted one.

[Stores have a checksum to detect corruption, but for efficiency this
code only reads the header of the file, waiting to verify the checksum
when the entire file is read later.]

BUG=279436

Review URL: https://chromiumcodereview.appspot.com/23546015