Properly forward queued touch events
If a touch-event is added to an empty TouchEventQueue and should not be
forwarded to the renderer, it is ack'ed to the client immediately.  However,
this ack may trigger follow-up events synchronously, which would be queued but
never forwarded to the renderer.  Properly forward the follow-up events when the
first event reaches an empty queue but was never forwarded.

BUG=230735

Review URL: https://chromiumcodereview.appspot.com/24312009