Make step name configurable for test runner tool.
This allows configuring a different step name used for the dashboard than the one used on the waterfall ui.

It is already used by the endure waterfall to name the test "endure/test_name" on the dashboard and "endure_test_name" on the waterfall.

The FP just got overwritten until now.

Review URL: https://chromiumcodereview.appspot.com/23621057