[Activity log] DOMActivityLogger::log now takes URL/title arguments
When calling into DOMActivityLogger::log, the current location and title
for the page should always be passed in--otherwise, when a
DOMActivityLogger is re-used across multiple pages stale data can be
logged.

This is the second in a set of related changes--the previous change,
https://codereview.chromium.org/23817008/, updates the blink code.  This
change is intended to be committed after that change lands in blink.

BUG=294500

Review URL: https://chromiumcodereview.appspot.com/24243011