The moment you've all been waiting for... the ninth happy number(1), 32.
The atomic number of germanium
The freezing point of water at sea level in degrees Fahrenheit
The number of completed, numbered piano sonatas by Ludwig van Beethoven
In chess, the total number of black squares on the board, the total number of white squares, and the total number of pieces (black and white) at the beginning of the game.
The number of teeth of a full set of teeth in an adult human, including wisdom teeth
The ball used in association football is typically made with 32 panels of leather or synthetic material.
The number of teams in the National Football League.
ASCII and Unicode code point for space " "
July 30th, 1932 - Walt Disney's Flowers and Trees, the first animated cartoon to be presented in full Technicolor, premieres in Los Angeles, California

Both Chrome for Windows and Mac are 32-bits (for now...)

Source(s)
(1) - http://en.wikipedia.org/wiki/Happy_number
http://en.wikipedia.org/wiki/32_(number)
http://en.wikipedia.org/wiki/1932