Recover corrupt Favicon databases.
Add sql::Recovery support to ThumbnailDatabase.  Rearrange the schema
setup to allow sharing with the recovery code.  Unit test a few
corruption cases.

Add sql::Recovery::Rollback() to allow transitioning code to monitor
failures for diagnostic purposes.  It is possible/likely that that
function can be removed if failure cases are infrequent enough (in
which case it with Unrecoverable()).

Expose sql::Connection::ShouldIgnore() to allow the error callback to
support sql::ScopedErrorIgnorer for testing.

BUG=240396

Review URL: https://chromiumcodereview.appspot.com/23533038