Remove passing of --revision for gclient on WebRTC ToT builds
WebRTC has a couple of builds that combines a Chromium checkout
with ToT WebRTC. These checkouts shouldn't specify the --revision
argument to gclient, as:
- it produces an error (with no impact though) for our 
  Chromium WebRTC FYI bots in 
  http://build.chromium.org/p/chromium.webrtc.fyi/waterfall
- causes the Chromium checkout get synced with the WebRTC
  revision the source poller has detected for our Android bots
  in http://build.chromium.org/p/client.webrtc/waterfall
  (started breaking after https://codereview.chromium.org/24315005)

With the changes made in
http://src.chromium.org/viewvc/chrome?revision=221758&view=revision
we can now start using the variable $$WEBRTC_REV$$ in the glient
spec since it will get replaced with the WebRTC revision.

Using this, we will now also be able to perform builds of older revisions
by entering the number in the revision field when forcing a build.

I also configured the 'no_gclient_revision' factory property to avoid
getting the --revision flag passed.

TEST=local runs with client.webrtc and chromium.webrtc.fyi waterfalls
with inspections of the gclient commands generated in the update steps.
BUG=none

Review URL: https://chromiumcodereview.appspot.com/24438007