Reduce browser_tests sharding to 5.
This is to increase throughput, so that the constant overhead is lower compared
to the execution. We are underprovisionned on Windows and they are the slowest
ones so we need to increase throughput. In practice we may want to use
OS-specific shard numbers but it not supported at the moment.

R=stip@chromium.org
BUG=

Review URL: https://chromiumcodereview.appspot.com/24242011