media: Implementing Opus Seek Pre Roll
Opus has the idea of a rolling key frame (i.e.) whenever there is
a discontinuity, certain number of samples has to be thrown away
before the decoded data is valid. The number of samples to be
thrown away can be found in the Seek Pre Roll element of the WebM
container. This CL makes use of that value to throw away those
samples after decoding whenever a seek happens and adjusts the
timestamp accordingly.

TEST=media unittests still pass

Review URL: https://codereview.chromium.org/23444076