Add Content-Length header to component extension url responses
When investigating the bug referenced below, we found that playing audio via
an <audio> tag from a file inside an extension worked if the extension was
installed normally, but not if it was a component extension. It turns out that
sending a Content-Length header makes this work properly.

BUG=284815

Review URL: https://codereview.chromium.org/23600029