Add support for zip+extract build based on a revision directory.
This CL adds a general-purpose way to specify another directory
than src/ to be used to specify the revision of the build archive.

Background:
WebRTC has several build configurations where a Chromium checkout
is used combined with ToT WebRTC. In these cases we want to number
the build archives created based on the WebRTC revision instead
of the Chromium revision.
Since the current implementation only looks for the revision in
the --src-dir, we need to be able to specify another directory,
similar to how WebKit does it. In our case it's the revision of
the third_party/webrtc folder that is interesting to use.

I realize this adds undesired boilerplate code, but I think WebKit
could probably use this instead of their webkit_dir factory
property. Unfortunately I think it's too risky to do that change in
this CL, since there may be many things depending on the existing
naming of their archives that is decided by
slave_utils.GetZipFileNames.

TEST=verified the correct file is used in the zip_build and
extract_build steps for local builds.
BUG=none

Review URL: https://chromiumcodereview.appspot.com/24586002