ChildProcessSecurityPolicy: Remove GrantCreateWriteFile.
The only user of GrantCreateWriteFile was the 'Save' file chooser. I figure if we're allowing the process to overwrite the file, it's harmless to allow it to also read the file.

This allows us to remove one permission from CPSP and simplify the code.

This is part of Step 2 of this refactoring plan:
https://docs.google.com/a/google.com/document/d/1QGkGWuwgSuaRqovz4wyb0upqPKDVsgYOFKt44E7gmOE/edit?usp=sharing

BUG=262142,263150

Review URL: https://codereview.chromium.org/24406003