sync: Introduce ModelNeutralWriteTransactions
Introduce ModelNeutralWriteTransactions that allows changes to entries
only if those changes do not need to be sent back to the model.  In
other words, they have write access to almost everything except
SPECIFICS, UNIQUE_POSITION, PARENT_ID, and other similar fields that
need to be reported back to the model to keep it in sync.

We enforce this restriction by allowing the construction of
ModelNeutralMutableEntries from these transactions, but requiring a full
WriteTransaction to construct regular MutableEntries.

To make it easier to write generic functions, this change also
introduces a common base class called BaseWriteTransaction.  This class
is a parent of both WriteTransaction and ModelNeutralWriteTransaction,
so it can be used in contexts where we're willing to accept either.
This is similar to the way BaseTransaction currently enables generic
functions to work with either ReadTransactions or WriteTransactions.

This patch does not actually instantiate any
ModelNeutralWriteTransactions.  The work of converting existing code to
use this new function will be started in future CLs.

BUG=284672

Review URL: https://codereview.chromium.org/24435002