[chromedriver] Handle the case that a command closes the window.
In some scenarios, like clicking a button or a link that closes the window, chrome will close the websocket connection between chromedriver and the devtools so fast that chromedriver won't get chance to receive any response for devtool commands and even Inspector.detached event.
So for a session command, if we get a Disconnected error and the window is already closed, we return OK instead.

Java acceptance test case: JavascriptEnabledDriverTest.testShouldBeAbleToClickALinkThatClosesAWindow

BUG=none

Review URL: https://codereview.chromium.org/24655002