Telemetry / Android: Fix branch number for Android WebView.
A recent change in Android WebView was setting the branch number in the
user agent string to 0. This was causing a crash in the
chrome_branch_number getter.
This change makes Android WebView telemetry scripts falling back to be
identify the WebViewShell as branch number 1025.

BUG=251008
NOTRY=true
R=bulach@chromium.org

Review URL: https://codereview.chromium.org/24641004