Fixes crasher when closing app windows while printing.
The ShellWindow acts as the delegate of a WebContentsModalDialogManager
for modal panels like the print preview panel. This change clears this
delegation relationship when the ShellWindow is closed by the user,
preventing a crash caused by the WebContentsModalDialogManager accessing
its ShellWindow delegate after ~ShellWindow is called.

There is a manual test case for the crasher in the bug report.

BUG=297179

Review URL: https://codereview.chromium.org/24363008