Read PNaCl fingerprint from disk.
This change is redundant given the pending refactoring of the PNaCl
installer, but that change is blocked on resolving the on-demand
feature. In the meantime, we want to read the read the diff update
fingerprint from disk to avoid distributing the wrong delta blobs.

BUG=245318

Review URL: https://codereview.chromium.org/24532003