cc: Add shared quad states correctly in unit tests.
When we add quads for tests we must do so in a valid way. We were
adding quads for the same SharedQuadState with quads for a different
SharedQuadState in between, but that's totally uncool.

Add a third SharedQuadState so that quads refer to SharedQuadStates
in a contiguous way as we would like in reality.

R=piman
BUG=263069

Review URL: https://codereview.chromium.org/24596005