cc: Fix the LayerTreeHostContextTestDontUseLostResources test.
The test was not drawing anything since it did not cause damage after
the first frame, and this meant that it was never losing its context
at all.

Also it was not setting up the capabilities to use io surfaces early
enough, so I started hitting asserts.

Also also, it was creating RenderPasses with ids reversed, so the
child render pass would never be used.

R=piman
BUG=263069

Review URL: https://codereview.chromium.org/24640003